% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy.R
\docType{data}
\name{epoxy}
\alias{epoxy}
\title{Epoxy Coating Corrosion Data}
\format{
A data frame with 12 rows and 3 columns:
\describe{
  \item{SYSTEM}{Factor indicating the epoxy coating system (levels 1–4)}
  \item{EXP-TIME}{Factor indicating the exposure time in hours (levels 1, 60, 120)}
  \item{CORRATE}{Numeric corrosion rate measured for each panel}
}
}
\source{
https://www.amazon.com/Statistics-Engineering-Sciences-William-Mendenhall/dp/1498728855
}
\usage{
epoxy
}
\description{
A dataset containing corrosion rate measurements for four epoxy coating systems exposed to water for varying durations. The data is used to evaluate the effect of coating type and exposure time on corrosion performance.
}
\keyword{datasets}
