% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrebase.R
\name{myrebase}
\alias{myrebase}
\title{Relevel Factors in Epoxy Dataset}
\usage{
myrebase(df, SYSTEM_base, EXP_base)
}
\arguments{
\item{df}{A data frame containing the epoxy data, with `SYSTEM` and `EXP-TIME` as factors or character vectors.}

\item{SYSTEM_base}{A value indicating the desired reference level for the `SYSTEM` factor.}

\item{EXP_base}{A value indicating the desired reference level for the `EXP-TIME` factor.}
}
\value{
A modified data frame with `SYSTEM` and `EXP-TIME` re-leveled to the specified baselines.
}
\description{
Reassigns the reference levels for the `SYSTEM` and `EXP-TIME` factors in the epoxy dataset.
This is useful for customizing baseline comparisons in linear models, especially when aligning with external software like SAS.
}
\examples{
df <- myrebase(df = epoxy, SYSTEM_base = 2, EXP_base = 60)
ylm <- lm(CORRATE ~ SYSTEM + `EXP-TIME`, data = df)
summary(ylm)
}
