% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epoxyapp.R
\name{run_epoxyapp}
\alias{run_epoxyapp}
\title{Launch the Epoxy App}
\usage{
run_epoxyapp()
}
\description{
Runs the interactive Shiny application that visualizes epoxy data.
}
\examples{
\dontrun{
run_epoxyapp()
}
}
